Refactor of packages for tests, Class for testing of messages
